[gui] Code cleanup
