--- src/emu/lazarus/config.json
@@ -1,1 +1,0 @@
-{ "rom" : "SYM.ROM", "cmap" : "drives/c/", "emap" : "drives/e/" }
